Dispatch updates even for non-drafts to backend
We need to do this because sets can be updated, and because we now use
SQLite to filter by mail sets, we need to ensure that SQLite has the
latest copy of the mail sets list for the mail.

Since we anyway already have to get the Mail entity and we don't need
to get details, this is free. Additionally, IndexedDb does not need any
changes as it always did mail set filtering inside of its respective
search facade.