Fix mail actions act on entire conversation with list grouping disabled
Introduced in c65f0f22
When viewing a conversation with conversation thread enabled but with
mail list grouping disabled, mail actions apply to the entire
conversation without exclusions (sent mails are supposed to be excluded
from move operations when acting on conversations). This occurs because
the entire conversation was used as actionable mails regardless of
whether list grouping was enabled or not.

Sent mails not being excluded from move operations was due to
`excludeMailSet` (passed to the server to exclude mails in a specific
folder from the being moved) being `null` when mail list grouping is
disabled.

To fix this, we use `getActionableMails()` which would return the mails
that the action will depend on. We renamed previous
`getActionableMails()` to `getResolvedMails()`.

Close #8886

Co-authored-by: ivk <ivk@tutao.de>