Setup release pipline for crypto-primitives.wasm
In order to make use of the crypto-primitives crate from the web client
we now generate wasm file using wasm-pack and included these steps into
the dev and webapp build process.

In order to test the behavior we also made use of the exported primitives from the Ed255519Facade

Also adds test for the Ed25519 exported functions.

tutadb#2102, tutadb#2099, tutadb#2098

Co-authored-by: hec <hec@tutao.de>
Co-authored-by: bedhub <bedhub@users.noreply.github.com>