Highlight additional fields (subject and sender/recipient)
We use the same SearchToken list from #8883 and extend it to be used
for the subject and mail viewer.

For the mail viewer, we create mark components with Mithril. This uses
a simple map() on the list returned by highlightTextInQuery function.

For the list view, since we use HTMLElement, we do what we did with the
mail body: create DOM elements, creating mark elements for the text that
needs highlighted. We use Text nodes for non-highlighted elements to
avoid malicious HTML injection, and we use textContent on highlighted
elements more-or-less does the same thing as making a Text node.

Closes #8939

Co-authored-by: paw <paw-hub@users.noreply.github.com>