Add keyboard shortcut to create new folder
Add a shortcut (ctrl-shift-N) to allow the user to create a
new folder without having to click on anything to open the
dialog.

Since a user may have access to more than one mail group, we
can disambiguate this by checking what folder they are looking
at (if any) and use this folder's owner group.

close #8922

Co-authored-by: paw <paw-hub@users.noreply.github.com>