Mail actions clear sticky mail and clear mail view
Previously mail view clearing was driven by the list state changing, but
in the case of opening a mail through a notification that is then moved
by an Inbox Rule we still want to view the mail even though it is no
longer in the list. Then doing an action on that mail would not clear
the mail view because the list state was not changing.

Co-authored-by: ivk <ivk@tutao.de>