Make overlay inert only when behind an open modal
The modal layer was initially assumed to always be on top of overlay.
However, we did not account for notification overlays, like the error
overlay for example.
So, in a situation where there's an error while a modal is open,
interacting with the error overlay is only possible once the modal is
closed.

An overlay should only be inert when modal is open and the overlay is
behind the modal (when overlay has a lower z-index than that of modal).

While this allows for notification overlays to be interacted with, they
are still not focusable, see #8995

Close #8767