Protect local identity keys from offline storage wiping
Previously the offline storage only contained some cached data, so it could be simply purged when it was causing trouble.

With the introduction of trusted identity keys that is no longer the case.

In this commit we make sure that the trusted identities are only deleted when intended, i.e., when removing the user's credentials, while still allowing the cached data to be cleared.
Note that a password change from a different device will cause the account to be logged out. If the new password is entered correctly, the trust database is preserved, but if the user chooses "logout" instead, then all data is wiped, including the trust database.

tutadb#2127