wip: fix server model info not being initialized
We observed clients that managed to get stuck with a missing server
model, which made login from stored credentials impossible and required
getting a new session.

The reason was that only the service executor, rest client and event bus
client cared to update the server model if they were missing.
However, the offline database also stores data in the server's model and
thus requires it to read any data. Since logging in will hit
the cache before requesting the server, a corrupted or missing server
model causes us to not be able to log in.

This commit gives the responsibility of acquiring the models to
ServerModelInfo.resolveServerTypeReference if they are missing or out
of date to avoid situations where we call resolveServerTypeReference
without first checking if the model is initialized.

Additionally, we show a dialog to the user if the models cannot be
initialized instead of an uncaught error.

We considered having the error handler try to repeat model
initialization, but since the original request already failed at that
point, getting back to a normal state would require an extra manual
retry.

TODO: remove the ApplicationTypesFacade from workerinterface

fix #9004

Co-authored-by: nig <nig@tutao.de>