Add filtering for mails inside of conversations
Make ConversationListModel display conversations that contain mails in
the folder that contain a given filter.

To avoid mails sorting in an unpredictable order as more mails are
loaded while the user is scrolling (and prevent having to load the
entire conversation!), the mail that is shown for the conversation is
the latest mail that matches the current filter, now.

Consequently, we now use the conversation ID for the list, although we
still sort using the displayed mail.

Closes #8989