Changes external calendars sync flow to not delete updated events
When deleting events that contains updates we must re-do all the
iteration over duplicated events to get their original id to delete,
and we also double the number of I/O operations since now we must
delete and recreate instead of just update.

This commit removes the updated events from the deletion list and
handle them in the classical way.