Have IndexedDbIndexer handle mail entity events
We removed directly handling entity events from MailIndexer, as we want
the events to be processed from offline storage, ensuring that events
are not missed. However, in doing so, events from IndexedDbIndexer are
ignored.

IndexedDbIndexer keeps track of its own batch ID, and we can simply call
the same functions that CustomMailCacheHandler would have called before
the batch ID is written.

Closes #9026

Co-authored-by: hrb <hrb-hub@users.noreply.github.com>