[mobile] better prompt for late contact permission
add a message dialog before the OS prompt for mobile contacts permission
make the second dialog in the permission flow more verbose

The contact permission can be requested in several ways, at least:
* from the setup wizard
* from the settings
* from the "Import" option in the Contact View

If the permission was denied before, the latter case did just
show the OS prompt as soon as the action was selected. The other two
had some explanation as to what was going on surrounding the prompt.