eventual consistent MailSetEntry ranges
We want to make sure that the MailSetEntry range is consistent after the
MailOfflineCleaner is completed. We accept that there might be a very short
time where the range stored in the offline db is inconsistent if the
MailOfflineCleaner is running in parallel to an operation that extends the
range into the past.