WIP: Create ContactSearchFacade
Surrounding a query in quotes has a different meaning between FTS5 and
IndexedDb searching. We don't want to expose which search model we are
using in this case, so we can just create a simple ContactSearchFacade
that handles any backend-specific details.

This simplifies a few details, like ContactModel originally needing a
lambda now only needing a ContactSearchFacade.

Closes #9088