fix external calendar events with organizer/attendees leading to resync
When performing a sync for an external calendar that has events
with an organizer or attendees, CalendarModel#eventHasSameFields
was not evaluating correctly. The deepEqual compare of
organizer / attendees aggregates lead to comparing aggregateIds (_ids)
and other technical fields that are not important when comparing the
actual event data. This commit fixes the problem, by comparing
stripped aggregates without the technical fields.