Fix showing own identity key in settings
All request that are exposed by the worker to the main thread must be async as they are executed async anyway when making the function call over the Main <-> Worker bridge.
In this case the calculateFingerprint function from the KeyVerificationFacade (which is part of the worker thread) has not been declared async.