Fix ConversationListModel being stuck when loading list offline
When loading of mails fails due to a network error the list model
tries to at least load the cached items. ConversationListModel was
repeatedly trying to load the items from offline because the start
id was not updated when loading items from the cache directly. This
would cause it to load and return the same items in the loop, without
ever throwing an offline error.

We fixed it by making sure the `lastFetchedMailSetEntryId` is set when
loading directly from cache which breaks up the loop.

Fix #9122

Co-authored-by: paw <paw-hub@users.noreply.github.com>