Fix Imported mails cannot be selected
Mails are sorted by received date in the MailListModel, but the encoded
date in the mailId is the date the mails were created in the server.
Therefore, the current stop condition using the mailId is not
sufficient to find out whether we have already loaded mails older than
the mail we're looking for. Fixed by limiting the pages to load, if the
targeted mail is within the pagelimit it will loadAndSelect otherwise
loading stop after the page limit.

Close: #8305

Co-authored-by: ivk <ivk@tutao.de>