Delete `metadata.lastUpdateTime` and recreate tables when purging
When migrating offline storage, `setCurrentOfflineSchemaVersion` is run
after `migrate`. Because of this, migrating would fail with a "no such
table: metadata" error if the migration purges storage, since purging
drops the metadata table.

We also don't recreate tables after migrating, which would result in the
same error if we try to write an entity after purging.

So instead of dropping the entire metadata table, we only delete
`metadata.lastUpdateTime`. We also create tables after migrating in
init.

Close #9129

Co-authored-by: paw <paw-hub@users.noreply.github.com>
Co-authored-by: ivk <ivk@tutao.de>