Fix birthday events for any time zone
The function getAllDayDateUTC assumes its params as UTC, basically swapping
its timezone to UTC but keeping the same day, month, year.
During a birthday creation timezones are applied on top of ISO String
(contact birthdayISO is considered UTC0).

When the user was in a negative timezone the local date would be one day behind.
This would cause the next call of getAllDayDateUTC (when setting endDate) to use
the past date as UTC and effectively set the endDate of the event as the same
day as the startDate.