Fix AssertNotNull failed missing mail in conversation
If we receive a mailSetEntry create event for a conversation mail that's
already loaded, the entry is inserted into the conversation mails list
without replacing, resulting in duplicates. This causes an inconsistent
state between mailMap and conversation mails when handling mailSetEntry
delete events, which happens because we assume that the conversation
mails list contains unique ids, and therefore we only remove once.

To fix this we simply make sure to avoid duplicates in conversation
mails list when inserting.

Close #9165

Co-authored-by: paw <paw-hub@users.noreply.github.com>
Co-authored-by: ivk <ivk@tutao.de>