Rework ConversationListModel
Make LoadedConversation into a class that manages its own state.

Have the conversation keep track of its mails, not the list model. We
can move a lot of code for mutating LoadedConversations into the
LoadedConversation itself and then call this from the list model.

Also fix how ConversationListModel does updates, since it is memoized
and may not update its arrays when expected.

Closes #9189
Closes #9196
Closes #9199

Co-authored-by: bir <bir@tutao.de>