Fix navigation race condition in SettingsView
Clicking on a folder updates browser URL immediately but mithril's
internal state is updated async. This could cause race condition when
we would click on a settings folder and then the admin folder loading
would finish and check against (now outdated) mithril route, switching
away from the folder we just selected.

We fixed it by resetting the nav target on selecting a settings folder
and adding a check for it.

Co-authored-by: ivk <ivk@tutao.de>