[tutanota] fix BadRequestError when adding aliases or users
MultipleMailAddressAvailabilityService now expects an optional
signupToken. It is only optional because the service can be called for
new customers during singups and for existing customers that want to
add another user or alias to their account.

The previous implementation of the SelectMailAddressForm always called
the signup version of the function.

fixes ee9b496ef1db41eb754c1ea8e585bc66efd5d5e8