[sdk] do not forward ignored_any to deserialize_any
Model changes to Group resulted in broken notification previews due to
deserialization errors. This commit ignores extra fields from the server models
when deserializing. Since we now write to the offlineDb using the server model,
this work-around is good enough as long as the SDK does not directly write to storage.

Co-authored-by: sug <sug@tutao.de>
Co-authored-by: charlag <charlag@tuta.io>