Handle key verification errors during verification process
Improve key verification dialog to:
* explain to the user what the error they are seeing might be;
* recommend a course of action;
* facilitate that course of action.

This includes allowing users to say they do not trust a key when
comparing fingerprints manually, and also to easily delete keys from
their local database.

TOFU and manually trusted keys are handled differently, because they
mean different things: if a key you previously manually verified does
not match the fingerprint you are currently being presented by your
contact, this is almost certainly a legitimate change and there is no
reason to keep the old key, but keeping the old key does not bring any
security risks with it. An error on a TOFU trusted key, though, might
mean an attack was taking place, and we strongly recommend the user to
discard that key and take the new one.

tuta#2197