fix concurrent loading of entityUpdates on mobile notification click
In case an entityUpdate, caused by an inbox rule or unread/read update,
happened while missedEntityUpdates were processed in the background,
it could happen that patches would have been applied twice. This commit
fixes the issue by also prefetching entityUpdates with patches or
instances on them, to ensure that the normal eventQueue handling
is not processing the entityUpdate a second time (because the update
is already in PrefetchStatus.Prefetched).

Co-authored-by: das <das@tutao.de>