Build crypto-primitives without debug symbols
The version of wasm-opt we were using (105) had a bug and the crate only
really worked with the `--dev` flag. This is not ideal, since it adds
some attack surface.

By adding binaryen (which provides wasm-opt) as an npm dependency with a
pinned version without this bug we can remove the flag and not break the
build.

The npm package is provided by the WebAssembly team.

tuta#2206