fix attachments not being decrypted correctly
When instances, such as tutanota/file(13) could not be decrypted
successfully, because the _ownerEncSessionKey == null, we were still
already putting them in the offline database when processing entity
updates. This lead to issues where attachments could never be decrypted
properly, even though the _ownerEncSessionKey is later added with a
patch entity update to the instance in the offline database. The
instance in the offline database would then be stored as decrypted
instance, however we assume that everything in the offline database
is already unencrypted. This commit fixes the issue.

Furthermore, the DefaultEntityRestCache#deleteFromCacheIfExists function
was broken, iterating over every character in an elementId,
instead of passing the actual elementId to the offline db sql query.