prefetch TutanotaProperties and do not prefetch ConversationEntry
We prefetch TutanotaProperties even though it is a element type as there
are a lot of missed entity updates for that type for some accounts and
not setting them to Prefetched would cause us to do a lot of GET
requests. We also do not prefetch ConversationEntries even if they have
more than one element anymore as we don't need them before displaying a
conversation and prefetching is not necessary.

Co-authored-by: das <das@tutao.de>