Cleanup; move Windows notification map to Windows notification factory
Having it in here means all of the Windows-specific code does not need
to be run on non-Windows platforms. For example, we do not really want
to keep track of notification IDs on Linux.

Admittedly, this might not be the best job for a factory, but at least
it means we don't need to worry about an id -> function map outside of
where it would be used.