do not put instance on the entity event in cache if the type is ignored
The DefaultEntityRestCache should not process with the entity events
for types we do not put in cache. We cannot filter them in the
filteredUpdates as some of the listeners could be using the updates to
update the UI even though the entities corresponding to that type are
not stored in the offline DB (e.g. Session type & LoginSettingsViewer).