Fix inconsistent mail counter badge
Counter fixup was happening before the list is fully loaded, resulting
in the fixup only reflecting part of the list, which would only be the
first page (100 mails) when the fixup happens after a list change but
before the user scrolls.

So we must wait for the list to fully load before attempting a fixup.

Close #9258

Co-authored-by: wrd <wrd@tutao.de>