Revert "[ios, ci] Use AdHoc entitlements for AdHoc builds"
This reverts commit 8c3bb931137ce605549bb31218a64b3639346c28.

We previously had to use separate AdHoc entitlements because the default
mail-client entitlement couldn't be included due to the AdHoc
provisioning profile not allowing it, causing the build to fail.

After we contacted Apple, they made it so our AdHoc profiles now allow
the mail-client entitlement, making this change unnecessary.