enable user satisfaction dialog for web & desktop
the user satisfaction dialog is shown based on the installation date of
the app.

on web, we don't have an "installation date", but we can add a timestamp
to the deviceConfig which stores the first time the app was opened in
that browser.

on desktop, the best we could do would be to inspect the current
executable file and using the creation date as the installation date.
This might be error prone since the behavior of such timestamps is
platform-dependent.
to avoid possible errors and the refactor of moving getInstallationDate
from MobileSystemFacade to CommonSystemFacade, we're using the same
method as on web and keep the native implementation for the mobile apps
since they have platform restrictions on how often to show the ratings
dialog base on the installation date.

Co-authored-by: yoy <yoy@tutao.de>