Fix delete action shown based on currentFolder regardless of mail folder
Permanent delete should only be allowed when the mail is in the current
folder and the current folder allows deletion (Trash/Spam). Otherwise,
the trash action is shown instead.

Previously we only checked the current folder. Meaning, if the current
folder is Trash/Spam, and we open a mail from notification, the delete
action was shown, even though the mail was neither in Trash nor Spam.

Close #9431

Co-authored-by: hrb-hub <hrb-hub@users.noreply.github.com>