Fix base64url.replace is not a function error when reactivating alias
The error is a result of an incorrect assumption that EntityUpdateData's
instanceListId is null for ElementEntities when processing entity
updates, when in fact instanceListId is an empty string.

Changing EntityUpdateData.instanceListId's type to NonEmptyString | null
and setting it to null for ElementEntity updates early should help
prevent bugs caused by such assumptions.

Close #9420

Co-authored-by: bir <bir@tutao.de>
Co-authored-by: ivk <ivk@tutao.de>