do not load encryption keys once for every attendee
tutadb#2396

When initializing an CalendarEventWhoModel for an existing event containing guest we only query encryption keys once for every attendee. Before it was querying the keys twice for the organizer, if the organizer was part of the attendees. This caused a "UNIQUE constraint failed" error for the IdentityKeyTrustDatabase because the recipient were resolved concurrently.