Ensure that the client has a defined state after signup
tutadb#2410, #9533

When signing up for a new account we need to create a session.
With the current setup we are also initializing some client models
which will remain in memory after closing the signup dialog.

We ensure now that we have a fresh client state by reloading the web view
as we don't have a proper method to re-initialize the MailLocator.

Co-authored-by: vaf <vaf@tutao.de>
Co-authored-by: sas <sas@tutao.de>