do not throw PatchOperationError when the path is not found
After a model change, the parsed instance on the offline database does
not necessarily match to the latest server model version. In that case,
the server might send entity updates containing information about
attributes which the offline database instance does not have. This does
not work when trying to patch that attribute on the client. In such
cases we now return null on PatchMerger to force a download from the
server to fetch the latest version of the instance.

Co-authored-by: Kinan <104761667+kibibytium@users.noreply.github.com>