Improve contrast fix
Instead of always forcing light theme, we can adapt the email to work in
dark mode.

We should still have a toggle in case our adaptation does not work for
a given email.

Additionally, the mail editor should have a toggle to toggle showing
mails with a light background or not.

Closes #9548