Fix throttle
This fixes a couple of problems with the throttle function.

First off, it wasn't actually being exported, so InboxRuleHandler was
importing it directly from its file. It's now correctly exported, and
it seems this was the *only* use of throttle.

Second off, lastArgs wasn't being saved, thus throttle wasn't actually
debouncing but just creating a new timeout for every invocation. For
example, if you called it 20 times with an interval of 5 seconds, it'd
just call it 20 times after a 5 second delay.

It now works like debounce() except that calling it does not restart the
timer. Hopefully this doesn't cause any spacebar heating (see xkcd 1172)
and InboxRuleHandler didn't actually depend on this behavior to work.