[build] Include crypto-primitives in workspace
crypto-primitives was not included in the cargo workspace, unlike the
rest of the crates. This caused issues with F-Droid builds as the
output directory is separate from the workspace. The crate also did not
benefit from shared dependency resolution and build caching.

We fixed it by including crypto-primitives in the workspace. We had to
work around the rustc profile issue: for wasm target we want to build
with size optimizations, but it's not possible to define it for one
target for one crate in the workspace. Instead, we've added a new
profile called `release-wasm` for the workspace and we specify it when
building `crypto-primitives` for wasm. To use the new profile we had to
change the version of `wasm-pack` as the one that's distributed via npm
does not include the `--profile` argument, even though the source code
for it is there. We use our own fork that is compiled at the install
time.

Co-authored-by: vis <vis@tutao.de>
Co-authored-by: paw <paw-hub@users.noreply.github.com>