Autosave drafts
Every five seconds while the user is typing, save a local copy of the
draft to ConfigurationDatabase.

If the draft has not been touched for five minutes, it gets uploaded to
the server.

If the user restarts the client with an autosaved draft, it will open
automatically. Only one draft can be autosaved at a time.

Additionally, if an entity update for a draft is received while the user
has that draft open for editing, saving will prompt a confirmation
dialog to overwrite the draft on the server, as this draft data may be
outdated.

Next, if the user attempts to open a new draft while there is an
autosaved draft, display a confirmation dialog to the user to delete
it.

And lastly, if the user has two separate drafts open on different
windows, we handle this by preventing saving over them.

Closes #4419