Do not autosave drafts if the body on the server changed
Fixes an oversight that can cause unintended data loss.

This oversight is caused by the fact that autosaving happens regardless
of if the draft was modified on another client.

Without this change, an autosave could (ironically) cause data loss if
the user has multiple clients and works on the draft on one client then
continues on another client. Because closing a draft automatically
saves the draft, it might save over the newer revision, causing data
loss.

For #4419 where the client will start periodically autosaving drafts,
the client will attempt to upload autosaved drafts on startup without
any input from the user. We only want an autosave to occur if this is
safe to do.

In the case that the user truly wants to save the draft, they can still
press CTRL-S (Command-S on macOS / iPadOS, etc.) to manually save the
draft while the draft editor is open.