Fixes event details page stuck
When an event from the widget, it might be that the event doesn't
exist anymore in case it has been deleted in another client.

This commit updates the onNewUrl method to handle cases where an event
might have been deleted from another client. If the initial sync has
completed and no event details are found, it now correctly assumes that
the event no longer exists.