Add isLightTheme/isDarkTheme, rename isDefaultLightTheme
isLightTheme checked if we were on the default light theme, not if the
actual theme was a light theme. To do that, we were using
`isColorLight(theme.content_bg)`. Except, we were still using the
isLightTheme function even in that case.

To remove confusion, we now have isDefaultLightTheme for when we want
to check if on the default light theme, and we have isLightTheme and
isDarkTheme to check if the current theme is a light or dark theme,
respectively.