refactor handling of birthday calendars so that they do not have fake group and groupInfo entities
We want to store the calendar color for birthday calendars on the server in order to synchronize it between devices and the widget.
There fore we should store the custom color on the UserSettingGroupRoot instance. We could use GroupSetting type for that
but we should use an existing group for it. The idea is that we use the contact group id for storing the color in the GroupSetting.
That way we don't need a model change.

In order to achieve that i want to refactor handling of birthday calendars in a way that we do not create fake group and groupInfo instance
for birthday calendars and rather make them their own type in the client. The approach also removes "clientOnly" calendar idea.
The only client only calendar we have right now is the birthday calendar and i would rather make this explicit so the code is more understandable.