fix native list unsubscribe post payload
When running our client on desktop or mobile, the list unsubscribe post
request is sent directly from the user device. However, in these cases,
we were setting a wrong post payload, which lead the request to fail
with some newsletter providers, such as GitHub. This commit also fixes,
that in case the post request fails, the unsubscribe fallback
action is executed correctly.

Co-authored-by: Kinan <104761667+kibibytium@users.noreply.github.com>