Add AppImage relaunch workaround
Adds a relaunch function to DesktopUtils and makes everything call that.
This works by calling the AppImage manually, and only if on Linux and
on a packaged AppImage.

We have to do this because app.relaunch() will not work on an AppImage,
and thus the program will simply close.

Close #8600

Co-authored-by: hrb <hrb-hub@users.noreply.github.com>