change handling of Entities with decryption errors when updating range
This commit makes a distinction between temporary and permanent crypto
errors. We do not write entities to the offline storage if they have
a SessionKeyNotFoundError, since this is a temporary error and is
expected to go away the next time the range is updated. We now do
cache entities with other errors, since we cannot do anything about
them.

Writing entities with permanent errors will not cause UI issues, since
the CryptoMapper replaces the affected fields with default values.

Co-authored-by: map <mpfau@users.noreply.github.com>
Co-authored-by: jhm <17314077+jomapp@users.noreply.github.com>
Co-authored-by: hrb-hub <181954414+hrb-hub@users.noreply.github.com>