Fix race conditions causing draft dialog conflict warnings
We want to avoid that an autosave is created while the user is saving.

It is not quite enough to check that the mail is sending before doing an
autosave, since you can also send while doing an autosave. As such, we
should additionally clear the autosave once the mail has been sent.

We also want to avoid saving if the mail has not actually been changed.

Closes #9735