Allow manually saving empty or unedited drafts
If the user presses Ctrl-S on an empty or unedited draft, then we should
actually save. This might be useful if the user is trying to create a
reply, for example, but may want to write the email later, and they do
not want to forget.

To do this, we can remove the hasMailChanged() check, since everything
that calls save() that isn't manually saving calls hasMailChanged()
synchronously, anyway.