Adds conflicting events information to Event Banner
When receiving an event invitation, users may have more
than just one conflicting event, but we show only one
inside the time overview.

This commit adds an expandable panel to the event banner,
allowing users to expand it by clicking in the warning info
and quickly see which events, along the already displayed one,
are conflicting with the invitation.

Closes #9726