Use offline storage for autosaved drafts
We want to use offline storage (SQLite) for autosaved drafts as it will
allow the locally saved draft to be accessible when offline. This means
we refactor ConfigurationDatabase.

We still want to use the same encoding methods, as it provides a simple
way to put in data without having to worry about making some complex
schema.

Closes #9743

Co-authored-by: ivk <ivk@tutao.de>