fix race condition when requesting identity keys for the first time
when putting the same mail address into different recipient fields and saving a draft there is an error when re-opening the draft later on a client that does not yet have the recipient's identity key. the identity key is then requested multiple times from the server in parallel, and we try to insert it multiple times into the tofu db. this failed. now we handle this as a noop if the row is exactly the same, we insert the new row if non exists for the primary key and we fail if there is a row for the primary key but it is not the same.
 this makes sure we do not accidentally override a key while handling the race condition gracefully.

tuta#2563