[iOS] Fix cannot change local storage range date
The date's value was being set on every redraw, and since we handle date
selection onfocusout on iOS, this would result in date having an
outdated value if a redraw (triggered by progress monitor for example)
happens between input and focusout events.

To mitigate this, we only set the date's value `oncreate`.

Close #9642

Co-authored-by: hrb-hub <hrb-hub@users.noreply.github.com>