Implements the new CalendarViewComponent and modularize the components
This commit extracts the CalendarView to its own component and break it in small pieces such as Header, TimeColumn, etc.

It rewrites the existing Multi-day Calendar View that previously
used absolute positioning to layout events, changing it to a more "Native"
approach using CSS Grid and moving the responsibility to deal with sizes
to the browser itself.

It is also expected to render events faster, since it does less floating point
math operations and to have less layout shifting when creating, editing and
deleting events.

Co-authored-by: and <and@tutao.de>